correcting conflicts when merging
